#29: Restructure packages and files using the feature-based project structure.
