+++ src/main/java/org/weathertrack/input/model/UserInputService.java
@@ -1,0 +1,5 @@
+package org.weathertrack.input.model;
+
+public interface UserInputService {
+	String getUserInput(String message);
+}
