--- src/main/java/org/weathertrack/service/WeatherDataProvider.java
@@ -1,7 +1,0 @@
-package org.weathertrack.service;
-
-import org.weathertrack.model.WeatherData;
-
-public interface WeatherDataProvider {
-	WeatherData getWeatherData(String city);
-}
