--- src/main/java/org/weathertrack/service/input/UserInputService.java
@@ -1,5 +1,0 @@
-package org.weathertrack.service.input;
-
-public interface UserInputService {
-	String getUserInput(String message);
-}
