--- src/main/java/org/weathertrack/service/validation/InputValidationExceptionMessage.java
@@ -1,9 +1,0 @@
-package org.weathertrack.service.validation;
-
-public class InputValidationExceptionMessage {
-	public static final String USER_INPUT_IS_NULL =
-			"User input is null!";
-
-	private InputValidationExceptionMessage() {
-	}
-}
