--- src/main/java/org/weathertrack/service/validation/InputValidator.java
@@ -1,9 +1,0 @@
-package org.weathertrack.service.validation;
-
-import org.weathertrack.service.validation.model.InputValidationResult;
-
-public interface InputValidator {
-	InputValidationResult validateCityNameInput(String userCityInput);
-
-	InputValidationResult validateMenuEntryInput(int userMenuEntry);
-}
