+++ src/main/java/org/weathertrack/ui/model/UserInterface.java
@@ -1,0 +1,13 @@
+package org.weathertrack.ui.model;
+
+import org.weathertrack.weather.model.WeatherData;
+
+import java.util.List;
+
+public interface UserInterface {
+	String getCityNameFromUser();
+
+	void printCitiesWithSameName(List<String> citiesWithSameName);
+
+	void printWeather(WeatherData weatherData);
+}
