+++ src/main/java/org/weathertrack/validation/model/InputValidationExceptionMessage.java
@@ -1,0 +1,9 @@
+package org.weathertrack.validation.model;
+
+public class InputValidationExceptionMessage {
+	public static final String USER_INPUT_IS_NULL =
+			"User input is null!";
+
+	private InputValidationExceptionMessage() {
+	}
+}
