+++ src/main/java/org/weathertrack/validation/model/InputValidationResult.java
@@ -1,0 +1,24 @@
+package org.weathertrack.validation.model;
+
+public class InputValidationResult {
+	private final boolean valid;
+	private final String validationMessage;
+
+	public InputValidationResult(boolean valid, String validationMessage) {
+		this.valid = valid;
+		this.validationMessage = validationMessage;
+	}
+
+	public InputValidationResult(boolean valid) {
+		this.valid = valid;
+		this.validationMessage = null;
+	}
+
+	public boolean isValid() {
+		return valid;
+	}
+
+	public String getValidationMessage() {
+		return validationMessage;
+	}
+}
