+++ src/main/java/org/weathertrack/validation/model/InputValidator.java
@@ -1,0 +1,7 @@
+package org.weathertrack.validation.model;
+
+public interface InputValidator {
+	InputValidationResult validateCityNameInput(String userCityInput);
+
+	InputValidationResult validateMenuEntryInput(int userMenuEntry);
+}
