+++ src/main/java/org/weathertrack/weather/model/WeatherDataProvider.java
@@ -1,0 +1,5 @@
+package org.weathertrack.weather.model;
+
+public interface WeatherDataProvider {
+	WeatherData getWeatherData(String city);
+}
